requirements/base.txt
Django==1.11.25
pytz==2019.3
MySQL-python==1.2.5
djangorestframework==3.9.4

